+++ tests/filters/test_html.py
@@ -61,8 +61,68 @@ class TestHTML(util.PluginTestCase):
         self.assertEqual(sorted(bad_words), words)
 
 
+class TestCSSEscapes(util.PluginTestCase):
+    """Test CSS escapes."""
+
+    def setup_fs(self):
+        """Setup file system."""
+
+        template = self.dedent(
+            """
+            <!DOCTYPE html>
+            <html>
+            <head>
+              <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+            </head>
+            <body>
+              <h1>A contrived example</h1>
+              <svg viewBox="0 0 20 32" class="icon icon-1" data1="1-20-40-50">
+                <use xlink:href="images/sprites.svg#icon-undo">aaaa</use>
+              </svg>
+              <svg viewBox="0 0 30 32" class="icon icon-2" data1="1-20-60-50">
+                <use xlink:href:colon="images/sprites.svg#icon-redo">bbbb</use>
+              </svg>
+              <svg viewBox="0 0 40 32" class="icon icon-3" data="70-20-30-50">
+                <use xlink:href="images/sprites.svg#icon-forward">cccc</use>
+              </svg>
+              <svg viewBox="0 0 50 32" class="icon icon-4" data="100-20-80-50">
+                <use xlink:href="other/sprites.svg#icon-reply">dddd</use>
+              </svg>
+            </body>
+            </html>
+            """
+        )
+
+        self.mktemp('test.txt', template, 'utf-8')
+
+    def test_css_escapes(self):
+        """Test css escapes."""
+
+        config = self.dedent(
+            """
+            matrix:
+            - name: html5_css_escapes
+              sources:
+              - '{}/**/*.txt'
+              aspell:
+                lang: en
+              pipeline:
+              - pyspelling.filters.html:
+                  mode: html5
+                  namespaces:
+                    xlink: http://www.w3.org/1999/xlink
+                  ignores:
+                  - '[xlink|href\\:colon^=images]'
+                  - '[xlink|\\68r\\65 f^=other]'
+            """
+        ).format(self.tempdir)
+        self.mktemp('.html5.yml', config, 'utf-8')
+        words = self.spellcheck('.html5.yml')
+        self.assertEqual(sorted(['aaaa', 'cccc']), words)
+
+
 class TestHtml5AttrNamespace(util.PluginTestCase):
-    """Test HTML plugin HTML5 namespace logic."""
+    """Test HTML plugin HTML5 attribute and namespace logic."""
 
     def setup_fs(self):
         """Setup file system."""